bkk v1.0.5+4 (20201124)

+++ lib/features/routes/home/data/models/game_platform.dart
@@ -128,5 +128,5 @@ extension GamePlatformEntityExtension on GamePlatformEntity {
     }
   }
 
-  bool isLongText(int limit) => label.countLength > limit;
+  bool isLongText(double limit) => label.countLength > limit;
 }
